pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "labelcheck"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "David Rickett", email = "david@va7eex.ca" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pyyaml"
]

[project.urls]
Documentation = "https://github.com/unknown/init#readme"
Issues = "https://github.com/unknown/init/issues"
Source = "https://github.com/unknown/init"

[tool.hatch.version]
path = "src/labelcheck/__about__.py"


[tool.pytest.ini_options]
pythonpath = ["src"]
minversion = "7.0"
addopts = "-ra -q --cov --cov-report=term --cov-report=xml"
testpaths = [
    "tests",
    "integration",
]

[tool.black]
target-version = ["py37"]
line-length = 120

[tool.isort]
profile = "black"

[tool.coverage.run]
source_pkgs = ["labelcheck", "tests"]
branch = true
parallel = true
omit = [
  "src/labelcheck/__about__.py",
  "src/labelcheck/__init__.py",
  "src/labelcheck/__main__.py",
  "**/__init__.py"
]

[tool.coverage.paths]
init = ["src/labelcheck", "*/labelcheck/src/labelcheck"]
tests = ["tests", "*/labelcheck/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
